# Make PrefixContext contain a varname rather than symbol